==> codes/01 Fitting Dataset.ipynb <==
"""
Jupyter Notebook file which takes in raw csv file obtained from Kaggle 
and preprocesses it by extracting relevant and useful hotel information 
and reviews and consequently storing them in their own respective files.

"""

# Import Libraries
import pandas as pd
import pickle as pk

# given an address from the review, function returns the city the hotel is located in
def city_identifier(address):
    address = address.split()
    if address[-1] == 'Kingdom':
        return 'London, United Kingdom'
    elif address[-1] == 'Austria':
        return 'Vienna, Austria'
    elif address[-1] == 'Netherlands':
        return 'Amsterdam, Netherlands'
    elif address[-1] == 'France':
        return 'Paris, France'
    elif address[-1] == 'Spain':
        return 'Barcelona, Spain'
    elif address[-1] == 'Italy':
        return 'Milan, Italy'
    return 'Another City'

# Read Data
raw = pd.read_csv('Stored Data\Hotel_Reviews_Sample.csv')
raw = raw.dropna(how="all")

# Create new column City and populate it by passing each address thru city_identifier by lambda
raw['City'] = raw['Hotel_Address'].apply(lambda x: city_identifier(x))

# Manually added latitute and longitude for hotels missing coordinates
adjusted  = pd.read_csv('Stored Data\Adjusted_Hotel_Info.csv')

# Create a smaller table with only columns that are useful
hotels = adjusted[['Hotel', 'Address', 'Score', 'Latitude', 'Longitude']]

# Populate City column in new table 'hotels', maintain indexes after removals
hotels['City'] = hotels['Address'].apply(lambda x: city_identifier(x))
hotels = hotels.reset_index(drop=True)

# Concatenate positive and negative review in one column 
pos = raw[['Hotel_Name', 'City', 'Positive_Review', 'Review_Date', 'Reviewer_Score']]
pos.columns = ['Hotel', 'City', 'Reviews', 'Date', 'Score']
neg = raw[['Hotel_Name', 'City', 'Negative_Review', 'Review_Date', 'Reviewer_Score']]
neg.columns = ['Hotel', 'City','Reviews', 'Date', 'Score']
reviews = pd.concat([pos, neg])

# Store the hotel information in its own pickle file
outfile = open("Stored Data/sample_hotels.pickle", "wb")
pk.dump(hotels, outfile)
outfile.close()

# Store the reviews information in its own pickle file
outfile = open("Stored Data/sample_reviews.pickle", "wb")
pk.dump(reviews, outfile)
outfile.close()

print(len(reviews))
print(len(hotels))
